Moved BaseFilterTestCase into main lib.
Makes it easier to write tests for custom filters.
